Change SwingTrajectoryParameters to not be abstract and update the default values to remove the ZuluSwingTrajectoryParameters.java
